requirements.txt
aiogram==3.17.0
aiohttp==3.11.12
python-dotenv==1.0.1
pytz==2025.1